Separate stream types and 'CostRose' into their own modules
