+++ plutus-core/plutus-core/src/PlutusCore/Builtin/Polymorphism.hs
@@ -20,7 +20,7 @@ import PlutusPrelude
 
 import PlutusCore.Builtin.HasConstant
 import PlutusCore.Core
-import PlutusCore.Evaluation.Machine.ExMemory
+import PlutusCore.Evaluation.Machine.ExMemoryUsage
 
 import Data.Kind qualified as GHC (Type)
 import GHC.Ix
