+++ plutus-core/plutus-core/src/PlutusCore/Evaluation/Machine/Ck.hs
@@ -31,7 +31,7 @@ import PlutusPrelude
 import PlutusCore.Builtin
 import PlutusCore.Core
 import PlutusCore.Evaluation.Machine.Exception
-import PlutusCore.Evaluation.Machine.ExMemory
+import PlutusCore.Evaluation.Machine.ExMemoryUsage
 import PlutusCore.Evaluation.Result
 import PlutusCore.Name
 import PlutusCore.Pretty (PrettyConfigPlc, PrettyConst)
